requirements.txt
# 依存ライブラリリスト
# 音量制御関連
pycaw==20230119
comtypes==1.1.14

# キーボード入力検出関連
pynput==1.7.6

# UI関連
PySimpleGUI==4.60.4
Pillow==9.5.0  # 画像処理用